refactor(sdk-guard): split the query-parameter guard under the 600-line cap (#15187)
The signature-derived coverage added for #15187 took
repo_tests/sdk_request_url_test.py to 643 lines, past
scripts/check_python_file_size.py's 600-line cap -- it was never in
python_file_size_known_large.py's grandfathered ratchet, so it fails
outright rather than needing a ceiling bump. Splitting rather than
grandfathering newly-written code, per the ratchet's own intent.

repo_tests/sdk_request_shared.py: the pinned SDK_REQUESTS table, the
httpx MockTransport wire-capture harness (_record/_urls), the
per-request query-parameter comparison (_assert_sent_params_are_accepted),
_template_for and _RESOURCE_ATTRS -- single-sourced so the two test
files below cannot drift apart on what SDK_REQUESTS contains.

repo_tests/conftest.py: the route_query_params fixture (and the
SERVING_MODULES it mounts), which both test files need. It lives here
rather than in sdk_request_shared.py because a fixture pytest finds by
directory scanning needs no import in either test module; importing it
by name for reuse would be exactly the "unused import" the repo-wide
autoflake pre-commit hook (.pre-commit-config.yaml, unscoped to
repo_tests/) strips on the next commit.

repo_tests/sdk_request_signature_params_test.py: the #15187
signature-derived section moved intact -- _dummy_value, _forced_call,
signature_forced_requests, both coverage tests and the contrast-mutation
test -- importing the shared pieces above rather than duplicating them.

repo_tests/sdk_request_url_test.py: trimmed to the #15053 URL-matching
guard, the #15119 per-row query-parameter guard, and the base-URL/env
tests, now importing SDK_REQUESTS and friends instead of defining them.

Pure move: no assertion logic changed, no test deleted, no entry added
to python_file_size_known_large.py.

Line counts: sdk_request_url_test.py 317, sdk_request_shared.py 143,
conftest.py 92, sdk_request_signature_params_test.py 189 -- all under
the 600-line cap.
